increased performance of operator*( quat, quat) by 33%
this in effect increased performance of Xforms xform(quat,vec) and operator*(q,v) by 12% for
floating (32 and 64), and by 30% for integer types.

did this with return value optimization.  previously this func was using mult(), which needed a
temporary to prevent aliasing.  operator* has no aliasing problems, so could be optimized...


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk/GGT/modules/GMTL@209 6382ae23-ed94-47d5-a97d-470f0bc4535f