the last commit of this file accidently checked in something I was trying.
because it is faster, use conj instead of invert in xform(quat,vec) and
operator*(quat,vec).  I've run tests, and using conj or invert both yield
the same results (cool), so use the faster one.

somehow I didn't trust my previous commit where I said to use conj,
but I've retested it, and I was right.  Now there is a test in the
test suite for it (to do the raw calculation of quat*vec using conj or
invert comparing if they equal, which they should.)


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk@233 6382ae23-ed94-47d5-a97d-470f0bc4535f