for both pnt and vec, divide the W param from the other elts, if it is non-zero...  this adds a yukky conditional... :(
also, this whole mat44 * pnt3 is a mess, and seems like it might be less
optimal than mat44 * pnt4.  in any case it is screwing up any chances for
having a mat34 * pnt4, that returns the correct value of pnt3

pnt3 = mat34 * pnt4  is correct matrix math, right?

well, as it is written now, we can only do:
pnt4 = mat34 * pnt4  (not correct matrix math)

cut currently xform doesn't support this... :(  so those tests have been
commented out.  this is probably because I don't know how to code it using
the generic methods I've been doing so far...

on second thought, because it is a transform, maybe we don't _want_ it to
change the size of the Vec or Pnt???  in that case, we may be good...

any help?


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk/GGT/modules/GMTL@248 6382ae23-ed94-47d5-a97d-470f0bc4535f