Bug fixed:	setDirCos and makeDirCos computed the incorrect matrix
This fixes a bug where the direction cosine code was incorrectly computing the matrix.  It computed the inverse of the correct matrix.

This change also adds a test that checks for correct behavior of this method and removes a previous test that is invalid since it just checked
to see that the (faulty) implementation did exactly what the implementation did.  (ie. it didn't actually test for behavior, it just tested the
implementation).

 CVS: ----------------------------------------------------------------------


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk@712 6382ae23-ed94-47d5-a97d-470f0bc4535f