Unconfused/Unobfuscated the matrix documentation.
The problem was that the previous docs didn't discuss interface, rather they assumed that all C/C++ matrices were accessed [row][col], same as gmtl.
Also, it talked in terms of OpenGL, rather than explaining really how GMTL was memory mapped.

I've hopefully clarified this, and in addition tried to keep the old information with the new flavor of how things are accessed, and some API
agnostisity.

Hopefully this is more clear to everyone.


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk@800 6382ae23-ed94-47d5-a97d-470f0bc4535f