Allow GMTL to continue being used by owners of Visual C++ 7.0, which has many shortcomings in terms of full ISO C++ compliance.
This change makes the metaprogramming code for vectors added by Allen to
be used conditionally based on whether the compiler supports the
constructs used.  At this time, the only compiler I have used that has any
trouble with the new code is Visual C++ 7.0.

Since nearly all of the old code was still in the files (commented out),
this change is quite minimal.

Bumped version to 0.4.3.

Approved by:	Silence on ggt-devel


git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk@1004 6382ae23-ed94-47d5-a97d-470f0bc4535f